Something is happening...
no useful data yet (all commands just empty), and cannot close stream
