+++ Configurator/src/views/Blackbox.vue
@@ -5,9 +5,9 @@ import Settings from "@components/blackbox/Settings.vue";
 import { BBLog, TraceInGraph, Command, LogData, LogDataType } from "@utils/types";
 import { constrain, intToLeBytes, leBytesToInt, prefixZeros } from "@utils/utils";
 import { skipValues } from "@/utils/blackbox/other";
-import { MspFn, MspVersion } from "@utils/msp";
+import { MspFn, MspVersion } from "@/msp/protocol";
 import { useLogStore } from "@stores/logStore";
-import { addOnCommandHandler, addOnConnectHandler, removeOnCommandHandler, removeOnConnectHandler, sendCommand } from "@/communication/serial";
+import { addOnCommandHandler, addOnConnectHandler, removeOnCommandHandler, removeOnConnectHandler, sendCommand } from "@/msp/comm";
 import TracePlacer from "@components/blackbox/TracePlacer.vue";
 import { BB_ALL_FLAGS, BB_GEN_FLAGS } from "@/utils/blackbox/bbFlags";
 import { parseBlackbox } from "@/utils/blackbox/parsing";
